#16: refactor styles
* moved SASS files into separate directory
  * updated associated references (to spinner image)
