+++ elections.html
@@ -7,12 +7,12 @@
 <meta name="author" content="Heini Fagerlund">
 <meta name="copyright" content="Heini Fagerlund">
 <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
-<link href="dist/css/main-b55ff1215f84cc13d86b.css" rel="stylesheet"></head>
+<link href="dist/css/main-c5ef2fe7aa73bbf73ad7.css" rel="stylesheet"></head>
 
 <body>
 <div id="app" class="ec-results"></div>
 
 <footer><p>Copyright &copy; 2018 Heini Fagerlund</p></footer>
 
-<script type="text/javascript" src="dist/js/main-b55ff1215f84cc13d86b.js"></script></body>
+<script type="text/javascript" src="dist/js/main-c5ef2fe7aa73bbf73ad7.js"></script></body>
 </html>
