+++ src/css/spinner.scss
@@ -5,7 +5,7 @@ $black:#000;
 }
 
 .spinner{
-background:transparent url("../../../assets/img/spinner.gif") no-repeat 50% 0% /  contain;
+background:transparent url("../assets/img/spinner.gif") no-repeat 50% 0% /  contain;
   color:$black;
   display:block;
   height:50px;
