+++ src/css/totalvotes.scss
